Added additional functionality for showing all products.

+++ db/migrate/20141216164646_create_coupon_usages.rb
@@ -1,15 +1,21 @@
 class CreateCouponUsages < ActiveRecord::Migration
   def self.up
-    create_table :coupon_usages do |t|
-      t.integer :coupon_id
-      t.integer :user_id
-      t.integer :usage_count
+    if not ActiveRecord::Base.connection.table_exists? 'coupon_usages' then
 
-      t.timestamps
+      create_table :coupon_usages do |t|
+        t.integer :coupon_id
+        t.integer :user_id
+        t.integer :usage_count
+
+        t.timestamps
+      end
     end
+        
   end
 
   def self.down
-    drop_table :coupon_usages
+    if ActiveRecord::Base.connection.table_exists? 'coupon_usages' then
+      drop_table :coupon_usages
+    end
   end
 end
